+++ rushub/src/Dir.cpp
@@ -174,7 +174,7 @@ void Dir::pathForFile(const char * file, string & path) {
 		slash = strrchr(file, '\\');
 	}
 	if (slash) {
-		path = string(file, slash - file + 1);
+		path = string(file, static_cast<size_t> (slash - file + 1));
 	} else {
 		path = '.';
 	}
